move redacted map and optimise
